# swagger-documentation.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_registerUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Registration succesful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/login-local":{"post":{"operationId":"AuthController_login_local","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginLocalDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/status":{"get":{"operationId":"AuthController_login_session","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/verify/{token}":{"post":{"operationId":"AuthController_confirmEmail","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/resend-code":{"post":{"operationId":"AuthController_resendVerification","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/resend-forgot-password":{"post":{"operationId":"AuthController_resendForgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/reset-password":{"patch":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/change-password":{"patch":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/users":{"get":{"operationId":"UsersController_getAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/{id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}}},"info":{"title":"Nestjs API Starter Kit","description":"This is my API starter kit to build a NestJs Backend API. It comes with Authentication, Authorization and Email setup.","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"cookie":{"type":"apiKey","in":"cookie","name":"connect.sid"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"password":{"type":"string"},"passwordConfirm":{"type":"string"}},"required":["name","email","phoneNumber","password","passwordConfirm"]},"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"password":{"type":"string"},"phoneNumber":{"type":"string"},"isEmailVerified":{"type":"boolean"}},"required":["id","name","email","role","password","phoneNumber","isEmailVerified"]},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"janedoe@mail.com"},"password":{"type":"string","description":"Password in plain text","example":"Pass@1234"}},"required":["email","password"]},"LoginLocalDto":{"type":"object","properties":{"username":{"type":"string","description":"Email address of the user","example":"janedoe@mail.com"},"password":{"type":"string","description":"Password in plain text","example":"Pass@1234"}},"required":["username","password"]},"EmailDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"johndoe@mail.com"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"password":{"type":"string","description":"Password in plain text","example":"Pass@1234"},"passwordConfirm":{"type":"string","description":"Password in plain text","example":"Pass@1234"},"resetToken":{"type":"string","description":"Reset code sent to email in plain text","example":"adu0f9"}},"required":["password","passwordConfirm","resetToken"]},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string","description":"Password in plain text","example":"Pass@1234"},"newPassword":{"type":"string","description":"Password in plain text","example":"Pass@1234"},"newPasswordConfirm":{"type":"string","description":"Password in plain text","example":"Pass@1234"}},"required":["oldPassword","newPassword","newPasswordConfirm"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"password":{"type":"string"},"passwordConfirm":{"type":"string"}}}}}}